+++ x-pack/plugin/security/src/test/java/org/elasticsearch/xpack/security/authc/kerberos/support/KerberosTestCase.java
@@ -69,6 +69,7 @@ public abstract class KerberosTestCase extends ESTestCase {
         unsupportedLocaleLanguages.add("hi");
         unsupportedLocaleLanguages.add("uz");
         unsupportedLocaleLanguages.add("fa");
+        unsupportedLocaleLanguages.add("ks");
     }
 
     @BeforeClass
